+++ packages/protocol/contracts/layer1/devnet/DevnetInbox.sol
@@ -23,7 +23,7 @@ contract DevnetInbox is Inbox {
     ///   _RING_BUFFER_SIZE = (86400 * D) / 12 / P
     ///                     = (86400 * 2) / 12 / 6
     ///                     = 2400
-    uint64 private constant _RING_BUFFER_SIZE = 100;
+    uint48 private constant _RING_BUFFER_SIZE = 100;
 
     // ---------------------------------------------------------------
     // Constructor
